Refactor some components
Adjust all components to use shared
Refactor createStorageDialog to be in localUI hook
