+++ src/lib/v2/hooks/local/useLocalUI.tsx
@@ -22,6 +22,7 @@ import ExportProgressItem, {
   ExportProcedureData,
 } from '../../../../components/molecules/Export/ExportProgressItem'
 import ExportSettingsComponent from '../../../../components/molecules/Export/ExportSettingsComponent'
+import { useRouter } from '../../../router'
 
 export function useLocalUI() {
   const { openSideNavFolderItemRecursively } = useGeneralStatus()
@@ -36,8 +37,42 @@ export function useLocalUI() {
     trashNote,
     purgeNote,
     renameStorage,
+    createStorage,
   } = useDb()
   const { pushMessage } = useToast()
+  const { push } = useRouter()
+
+  const openCreateStorageDialog = useCallback(() => {
+    openModal(
+      <BasicInputFormLocal
+        defaultIcon={mdiFolderOutline}
+        defaultInputValue={''}
+        defaultEmoji={undefined}
+        placeholder='Workspace name'
+        submitButtonProps={{
+          label: 'Create Space',
+        }}
+        onSubmit={async (workspaceName: string) => {
+          if (workspaceName == '') {
+            pushMessage({
+              title: 'Cannot rename workspace',
+              description: 'Workspace name should not be empty.',
+            })
+            closeLastModal()
+            return
+          }
+          const storage = await createStorage(workspaceName)
+          // todo: [komediruzecki-30/05/2021] should also update to proper initial screen (open sidebar state etc)
+          push(`/app/storages/${storage.id}/notes`)
+          closeLastModal()
+        }}
+      />,
+      {
+        showCloseIcon: true,
+        title: 'Create a space',
+      }
+    )
+  }, [closeLastModal, createStorage, openModal, push, pushMessage])
 
   const openWorkspaceEditForm = useCallback(
     (workspace: NoteStorage) => {
@@ -388,6 +423,7 @@ export function useLocalUI() {
     // deleteWorkspace,
     deleteOrTrashNote: deleteOrArchiveDoc,
     exportDocuments,
+    openCreateStorageDialog,
   }
 }
